pyproject.toml
[project]
name = "mistral-mathpix"
version = "0.1.0"
description = "Desktop tray application for OCR using Mistral"
requires-python = ">=3.8"
dependencies = [
    "pillow",   # For image handling
    "mistralai",  # Mistral AI client
    "python-dotenv",  # For environment variables
    "pnglatex",  # For rendering LaTeX to PNG
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mistral_mathpix"]

[tool.ruff]
select = ["E", "F", "I"]
ignore = []
line-length = 88 
